requirements.txt
argh==0.26.2
asciimatics==1.10.0
atomicwrites==1.2.1
attrs==18.2.0
fake-factory==0.7.2
Faker==0.7.3
future==0.17.1
more-itertools==4.3.0
numpy==1.11.2
pathtools==0.1.2
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pyfiglet==0.7.6
pygtail==0.10.0
pyparsing==2.3.0
pytest==4.0.1
python-dateutil==2.7.5
pytz==2016.7
PyYAML==3.13
six==1.11.0
tzlocal==1.3
watchdog==0.9.0
wcwidth==0.1.7
